2016年 2月23日 星期二 16时15分34秒 CST

+++ recovery.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
 
-git pull
+git_pull_log=`git pull`
+
+if [ ${git_pull_log} == "Already up-to-date." ]
+	return
+fi
+
+git show
 
 echo  "press enter key to continue"
 read
+
 
 cp ./init/.lldbinit ~/
 cp ./init/.vimrc ~/
